added boilerplate and update old example test
